Added two XiongBot VIPs at different positions. Added two TestChaserBot chasers at different positions. Set up chaser positions for both VIPs to track. Simulated turns where both chasers and VIPs take actions.

+++ src/ICS4U_Summative/TestChaserBot.java
@@ -58,7 +58,7 @@ public class TestChaserBot extends BaseBot {
      * Helper method to get moves per turn
      * @return the number of moves this robot can make per turn
      */
-    private int getMovesPerTurn() {
+    private double getMovesPerTurn() {
         return super.getDodgeDifficulty(); // Using dodge difficulty as a proxy for moves per turn
     }
 }
